sync w/ upstream

+++ tasks/upload-product-and-stemcell/task.sh
@@ -1,32 +1,49 @@
-#!/bin/bash -e
+#!/bin/bash
 
-if [[ ! -z "$NO_PROXY" ]]; then
+set -eu
+
+if [[ -n "$NO_PROXY" ]]; then
   echo "$OM_IP $OPS_MGR_HOST" >> /etc/hosts
 fi
 
-PIVNET_CLI=`find ./pivnet-cli -name "*linux-amd64*"`
-chmod +x $PIVNET_CLI
+STEMCELL_VERSION=`cat ./pivnet-product/metadata.json | jq --raw-output '.Dependencies[] | select(.Release.Product.Name | contains("Stemcells")) | .Release.Version'`
 
-chmod +x tool-om/om-linux
-CMD=./tool-om/om-linux
+if [ -n "$STEMCELL_VERSION" ]; then
+  diagnostic_report=$(
+    om-linux \
+      --target https://$OPS_MGR_HOST \
+      --username $OPS_MGR_USR \
+      --password $OPS_MGR_PWD \
+      --skip-ssl-validation \
+      curl --silent --path "/api/v0/diagnostic_report"
+  )
+
+  stemcell=$(
+    echo $diagnostic_report |
+    jq \
+      --arg version "$STEMCELL_VERSION" \
+      --arg glob "$IAAS" \
+    '.stemcells[] | select(contains($version) and contains($glob))'
+  )
+
+  if [[ -z "$stemcell" ]]; then
+    echo "Downloading stemcell $STEMCELL_VERSION"
+    pivnet-cli login --api-token="$PIVNET_API_TOKEN"
+    pivnet-cli download-product-files -p stemcells -r $STEMCELL_VERSION -g "*${IAAS}*" --accept-eula
+
+    SC_FILE_PATH=`find ./ -name *.tgz`
+
+    if [ ! -f "$SC_FILE_PATH" ]; then
+      echo "Stemcell file not found!"
+      exit 1
+    fi
+
+    om-linux -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k upload-stemcell -s $SC_FILE_PATH
+
+    echo "Removing downloaded stemcell $STEMCELL_VERSION"
+    rm $SC_FILE_PATH
+  fi
+fi
 
 FILE_PATH=`find ./pivnet-product -name *.pivotal`
-
-STEMCELL_VERSION=`cat ./pivnet-product/metadata.json | jq '.Dependencies[] | select(.Release.Product.Name | contains("Stemcells")) | .Release.Version'`
-
-echo "Downloading stemcell $STEMCELL_VERSION"
-$PIVNET_CLI login --api-token="$PIVNET_API_TOKEN"
-$PIVNET_CLI download-product-files -p stemcells -r $STEMCELL_VERSION -g "*vsphere*" --accept-eula
-
-SC_FILE_PATH=`find ./ -name *.tgz`
-
-$CMD -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k upload-product -p $FILE_PATH
-
-$CMD -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k upload-stemcell -s $SC_FILE_PATH
-
-if [ ! -f "$SC_FILE_PATH" ]; then
-    echo "Stemcell file not found!"
-else
-  echo "Removing downloaded stemcell $STEMCELL_VERSION"
-  rm $SC_FILE_PATH
-fi
+om-linux -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k --request-timeout 3600 upload-product -p $FILE_PATH
